FIX: handle long long int on 64bit platforms
git-svn-id: https://www.nitrc.org/svn/cmtk/trunk@5382 42a5c34f-2066-0410-bec5-ba365beb4995

+++ core/libs/System/cmtkCommandLineItem.cxx
@@ -4,6 +4,8 @@
 //
 //  Copyright 2004-2009 SRI International
 //
+//  Copyright 2015 Google, Inc.
+//
 //  This file is part of the Computational Morphometry Toolkit.
 //
 //  http://www.nitrc.org/projects/cmtk/
@@ -29,6 +31,7 @@
 //  $LastChangedBy$
 //
 */
+
 #include "cmtkCommandLine.h"
 
 long int
@@ -36,7 +39,7 @@ cmtk::CommandLine::Item
 ::ConvertStrToLong( const char* str )
 {
   char* endptr;
-  const int value = strtol( str, &endptr, 0 );
+  const long int value = strtol( str, &endptr, 0 );
   if ( (endptr == str) || *endptr )
     {
     throw( Exception( "Option expects an integer argument" ) );
